Update blog rendering + add pagination display
Updated renderPosts - we now render posts based on the page number and the number of items per page. Each time the user adds a post, the blog posts container is cleared and the blog posts that relate to the current page number are rendered to the container.

Added a updatePaginationDisplay method that simply hides or displays pagination based on whether the number of blog posts exceeds the items per page count.