Update pagination functionality
Added various elements in HTML to accomodate pagination functionality. We have hard-coded prev and next "buttons" with a <ul> in between. Each time we render the pagination element, we clear this <ul> and add the required page numbers: 1, 2, 3 etc.

Styled lists so there's no text decoration or list style.

Added js to get the <ul> list between next and prev. Added an event listener to this list. When the user clicks on a number, we update the pageNum and call renderPosts.

Added a function, renderPagination, to clear the pagination list and add list items to it as needed.