Update styles and js to accommodate styles
Added css to highlight current page selection in pagination.

Updated paginationNumberList event listener. Each time a pagination list item is clicked, we remove the 'current-page' class from all elements and add it to the clicked item.

Updated renderPagination so it adds the 'current-page' class to the list item that has an id that matches the global pageNum variable.