Update prev,next functionality for pagination
Updated the updatePaginationDisplay to show/hide prev and next where the user is at the first or last page.

Added call to updatePaginationDisplay in the paginationNumberList event listener.

When the page number is 1, prev is hidden. When page number is greater than 1, prev is shown. When the user is viewing the last page, next is hidden. When the page number is less than the last page number, next is shown.